Add support for small format
